monitoreo/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
Pillow>=10.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
mysqlclient>=2.2.0
gunicorn>=21.0.0
whitenoise>=6.5.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
requests>=2.31.0